Address Andrew's comments
