+++ client/src/constants/analysis_constants.js
@@ -20,4 +20,38 @@ const COLOR_MAP = {
   positive: { label: 'Positive', color: 'green', background: LIGHT_GREEN },
 };
 
-export { COLOR_MAP, INPUT_LIST, INPUT_ROWS };
+const mockContentTerms = [
+  '{"name": "Donald Trump", "type": "PERSON", "salience": 0.61}',
+  '{"name": "Georgia", "type": "LOCATION", "salience": 0.2}',
+  '{"name": "re-election", "type": "EVENT", "salience": 0.1}',
+  '{"name": "help", "type": "OTHER", "salience": 0.08}',
+  '{"name": "Chip", "type": "PERSON", "salience": 0.01}',
+  '{"name": "2020", "type": "DATE", "salience": 0.0}',
+  '{"name": "$5", "type": "PRICE", "salience": 0.0}',
+  '{"name": "2020", "type": "NUMBER", "salience": 0.0}',
+  '{"name": "5", "type": "NUMBER", "salience": 0.0}',
+];
+
+const mockHeaderTerms = [
+  '{"name": "Georgia", "type": "LOCATION", "salience": 0.9}',
+  '{"name": "Make America Great Again", "type": "OTHER", "salience": 0.1}',
+  '{"name": "2020", "type": "DATE", "salience": 0.0}',
+  '{"name": "2020", "type": "NUMBER", "salience": 0.0}',
+];
+
+const mockHeaderScore = 0.4;
+const mockHeaderMagnitude = 0.7;
+
+const mockContentScore = -0.1;
+const mockContentMagnitude = 0.2;
+
+const MOCK_DATA = {
+  mockContentTerms,
+  mockHeaderTerms,
+  mockHeaderScore,
+  mockHeaderMagnitude,
+  mockContentScore,
+  mockContentMagnitude,
+};
+
+export { COLOR_MAP, INPUT_LIST, INPUT_ROWS, MOCK_DATA };
